:recycle: Bedre exports av utils
