+++ @navikt/core/react/src/loader/Loader.tsx
@@ -1,7 +1,6 @@
-import React, { forwardRef, SVGProps } from "react";
 import cl from "clsx";
-import { useId } from "../util/useId";
-import { omit } from "../util";
+import React, { forwardRef, SVGProps } from "react";
+import { omit, useId } from "../util";
 
 export interface LoaderProps extends SVGProps<SVGSVGElement> {
   /**
