+++ @navikt/core/react/src/util/omit.ts
@@ -1,0 +1,10 @@
+export const omit = (obj: object, props: string[]) =>
+  Object.entries(obj)
+    .filter(([key]) => !props.includes(key))
+    .reduce(
+      (obj, [key, value]) => ({
+        ...obj,
+        [key]: value,
+      }),
+      {}
+    );
