+++ aksel.nav.no/website/package.json
@@ -62,7 +62,7 @@
   "devDependencies": {
     "@axe-core/playwright": "^4.5.2",
     "@babel/core": "^7.19.0",
-    "@next/bundle-analyzer": "^12.1.0",
+    "@next/bundle-analyzer": "^13.5.0",
     "@next/eslint-plugin-next": "^12.2.0",
     "@playwright/test": "1.35.1",
     "@sanity/block-tools": "^3.16.2",
