+++ examples/next-appdir/next.config.js
@@ -1,4 +1,4 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {};
-
-module.exports = nextConfig;
+const withBundleAnalyzer = require("@next/bundle-analyzer")({
+  enabled: true,
+});
+module.exports = withBundleAnalyzer({});
