+++ examples/next-appdir/package.json
@@ -16,6 +16,7 @@
     "react-dom": "^18.0.0"
   },
   "devDependencies": {
+    "@next/bundle-analyzer": "^13.5.0",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "typescript": "^5.1.6"
